backend/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 数据库驱动
motor==3.5.1                    # MongoDB异步驱动 (兼容新版PyMongo)
pymongo==4.8.0                  # 指定兼容的PyMongo版本

# HTTP客户端
httpx==0.25.2                   # 异步HTTP客户端
aiofiles==23.2.1                # 异步文件操作

# 缓存
redis==5.0.1                    # Redis客户端
aioredis==2.0.1                 # 异步Redis客户端

# 数据验证和序列化
pydantic==2.5.0                 # 数据验证
pydantic-settings==2.1.0       # 设置管理

# 日志和监控
structlog==23.2.0               # 结构化日志
prometheus-client==0.19.0       # 监控指标

# 工具库
python-slugify==8.0.1           # URL slug生成
Pillow==10.1.0                  # 图像处理
pypinyin==0.49.0                # 中文转拼音
jieba==0.42.1                   # 中文分词

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 性能优化
orjson==3.9.10                  # 高性能JSON库
uvloop==0.19.0                  # 高性能事件循环（仅Linux）

# 安全
cryptography==41.0.7
